backend/requirements.txt
captcha==0.4
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.2
opencv-python-headless==4.7.0.72
pi==0.1.2
Pillow==9.4.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
